+++ src/touchable.tsx
@@ -61,7 +61,7 @@ interface CommonProps {
     disabled?: boolean;
     elementRef?: React.RefObject<HTMLButtonElement | HTMLAnchorElement | HTMLDivElement>;
     style?: React.CSSProperties;
-    trackingEvent?: TrackingEvent;
+    trackingEvent?: TrackingEvent | ReadonlyArray<TrackingEvent>;
     label?: string;
     'data-testid'?: string;
     'aria-checked'?: 'true' | 'false' | boolean;
@@ -125,6 +125,14 @@ const Touchable: React.FC<Props> = (props) => {
     const {texts, analytics, platformOverrides, Link} = useTheme();
     const classes = useStyles();
     const isClicked = React.useRef(false);
+    let trackingEvents: ReadonlyArray<TrackingEvent> = [];
+    if (props.trackingEvent) {
+        if (Array.isArray(props.trackingEvent)) {
+            trackingEvents = props.trackingEvent;
+        } else {
+            trackingEvents = [props.trackingEvent as TrackingEvent];
+        }
+    }
 
     const children = props.children;
 
@@ -165,8 +173,7 @@ const Touchable: React.FC<Props> = (props) => {
         return '';
     };
 
-    const trackEvent = () =>
-        props.trackingEvent ? analytics.logEvent(props.trackingEvent) : Promise.resolve();
+    const trackEvent = () => Promise.all(trackingEvents.map((event) => analytics.logEvent(event)));
 
     const trackOnce = (callback: () => void) => {
         if (isClicked.current) return;
@@ -180,7 +187,7 @@ const Touchable: React.FC<Props> = (props) => {
 
     const handleButtonClick = (event: React.MouseEvent<HTMLElement>) => {
         // synchronously execute handler when no tracking is needed
-        if (!props.trackingEvent) {
+        if (!trackingEvents.length) {
             onPress(event);
             return;
         }
@@ -189,7 +196,7 @@ const Touchable: React.FC<Props> = (props) => {
     };
 
     const handleHrefClick = (event: React.MouseEvent<HTMLElement>) => {
-        if (!props.trackingEvent) {
+        if (!trackingEvents.length) {
             return; // leave the browser handle the href
         }