+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGELOG
 
-## unreleased
+## Unreleased
 
-- New FormSearchField component
+- Support multiple tracking events
+- New `FormSearchField` component
 
 ## v.5.1.0 - 2020-08-12